requirements.txt
aiohttp
bs4
PyYaml
termcolor
tqdm